fix: gesture robot direction anywhere
Previously you can only gesture the next position of the robot from the field component,
which did dificult select next direction if you are near to any wall. However now the
gesture can be selected from any point of the view.

To make this possible, a new div tag was created into the Game component to wrap all the code, and
then listening to onMouseMove and onMouseUp events from there insteand of do it from the Field component.