fix: set rotation as admin without scrolling view
Previously when admin wants to rotate a robot, the view was scrolling at the same momment when the robot was rotating.
That made the rotation unfunctional as an admin.

This fix introduce new events to detect when the pointer is enter or leave the robot and in this way, block or allow the scroll in admin view