fix: set rotation of robot with gestures
Before this, as an admin you are just able to set the rotation using the mouse wheel, which represent an issue if the admin is using a mobile device.
Addressing those changes, the admin is going to be able to set the rotation similar than the user does on Game view.