feat(admin): reposition robots using long-press (#50)
Add a new behavior that allows the re-positioning of robots on the admin view by pressing on the robot for an amount of time bigger than .5 seconds. Below that time, the admin is going to be able to set the rotation of the robot by pressing on it, but abow that time the reposition of the robot is allowed. In this way, admin now is able to set new position and new rotation no matter the device they are using.

Co-authored-by: Markus Tacker <Markus.Tacker@NordicSemi.no>