fix: update gateway only when all teams are ready to fight
Right now the gateway is updated just after the user set the desired movement for a single robot, that makes the robot moves even when the 'fight' button havent been pressed.

The rules of the game said that the robot should move only when all the teams are ready to fight.

To accomplish that rule, the notification to the gateway have been removed when the user set the desired movement for a robot and insteand do it when all the teams are ready to fight